// src/interfaces/strapi.interface.ts
// Generated by https://quicktype.io

export interface IStrapiError {
    data:  null;
    error: Error;
}

export interface Error {
    status:  number;
    name:    string;
    message: string;
    details: Details;
}

export interface Details {
}



export interface StrapiResponse <T> {
    data: T;
    meta: Meta;
}

export interface Meta {
    pagination?: Pagination;
}

export interface Pagination {
    page:      number;
    pageSize:  number;
    pageCount: number;
    total:     number;
}