// src/interfaces/listaContactos.interfaces.ts
// Generated by https://quicktype.io

import { IUser } from "./auth.interfaces";
import { IContacto } from "./contactos.interfaces";

export interface IListaContacto {
    id:         number;
    attributes: IListaContactoAttributes;
}

export interface IListaContactoAttributes {
    nombre:      string;
    createdAt:   string;
    updatedAt:   string;
    publishedAt: string;
    contactos?: {
        data: IContacto[]
    }
    usuario: {
        data: IUser
    }
}